Moved generate_messages_file function to cli.py.  First commit on command_line_two branch
